A problem in which you must state how many times a number appears in an array.
I want to come back to this one and make it more efficient as it passes 4/5 of the HackerEarth tests, failing the third one on time taken. I understand that this way is very inefficient and want it to pass the final test. I'll use something like an ArrayList of a pair class to store each number and the occurrences. This will make reading values quicker.